2002-09-14  Martin Baulig  <martin@gnome.org>
	* MonoSymbolTable.cs: Use only CLS-compliant types here, bumped version
	number to 19.

	* MonoSymbolTableReader.cs: Removed, this is now in the debugger.

svn path=/trunk/mcs/; revision=7447

+++ mcs/class/Mono.CSharp.Debugger/MonoSymbolTableWriter.cs
@@ -47,24 +47,24 @@ namespace Mono.CSharp.Debugger
 			Hashtable sources = new Hashtable ();
 			int source_idx = 0;
 
-			ot.source_table_offset = (uint) bw.BaseStream.Position;
+			ot.source_table_offset = (int) bw.BaseStream.Position;
 			foreach (ISourceFile source in symwriter.Sources) {
 				if (sources.ContainsKey (source))
 					continue;
 
-				sources.Add (source, (uint) bw.BaseStream.Position);
+				sources.Add (source, (int) bw.BaseStream.Position);
 				bw.Write (source.FileName);
 			}
-			ot.source_table_size = (uint) bw.BaseStream.Position - ot.source_table_offset;
+			ot.source_table_size = (int) bw.BaseStream.Position - ot.source_table_offset;
 
 			//
 			// Write line number table
 			//
 			Hashtable methods = new Hashtable ();
 
-			uint address_table_size = 0;
+			int address_table_size = 0;
 
-			ot.line_number_table_offset = (uint) bw.BaseStream.Position;
+			ot.line_number_table_offset = (int) bw.BaseStream.Position;
 			foreach (ISourceMethod method in symwriter.Methods) {
 				if (method.Start == null || method.Start.Row == 0)
 					continue;
@@ -72,10 +72,10 @@ namespace Mono.CSharp.Debugger
 				int count = method.Lines.Length;
 				LineNumberEntry[] lines = new LineNumberEntry [count];
 
-				uint pos = (uint) bw.BaseStream.Position;
+				int pos = (int) bw.BaseStream.Position;
 
-				uint address_table_offset = address_table_size;
-				uint my_size = (uint) (MethodAddress.Size + count * 8);
+				int address_table_offset = address_table_size;
+				int my_size = (int) (MethodAddress.Size + count * 8);
 				address_table_size += my_size;
 
 				for (int i = 0; i < count; i++) {
@@ -84,30 +84,30 @@ namespace Mono.CSharp.Debugger
 				}
 
 				MethodEntry entry = new MethodEntry (
-					(uint) method.Token, (uint) sources [method.SourceFile],
+					(int) method.Token, (int) sources [method.SourceFile],
 					method.SourceFile.FileName, lines, pos, address_table_offset,
-					my_size, (uint) method.Start.Row, (uint) method.End.Row);
+					my_size, (int) method.Start.Row, (int) method.End.Row);
 
 				methods.Add (method, entry);
 
 			}
-			ot.line_number_table_size = (uint) bw.BaseStream.Position -
+			ot.line_number_table_size = (int) bw.BaseStream.Position -
 				ot.line_number_table_offset;
 
 			//
 			// Write method table
 			//
-			ot.method_count = (uint) methods.Count;
-			ot.method_table_offset = (uint) bw.BaseStream.Position;
+			ot.method_count = methods.Count;
+			ot.method_table_offset = (int) bw.BaseStream.Position;
 			foreach (MethodEntry entry in methods.Values)
 				entry.Write (bw);
-			ot.method_table_size = (uint) bw.BaseStream.Position -  ot.method_table_offset;
+			ot.method_table_size = (int) bw.BaseStream.Position -  ot.method_table_offset;
 
 			//
 			// Write offset table
 			//
 			ot.address_table_size = address_table_size;
-			ot.total_file_size = (uint) bw.BaseStream.Position;
+			ot.total_file_size = (int) bw.BaseStream.Position;
 			bw.Seek ((int) offset_table_offset, SeekOrigin.Begin);
 			ot.Write (bw);
 			bw.Seek (0, SeekOrigin.End);
